+++ LTPT_UngDungThiTracNghiem_DHKTPM18C/src/main/java/service/HocSinhService.java
@@ -4,6 +4,6 @@ import entities.HocSinh;
 
 import java.rmi.RemoteException;
 
-public interface HocSinhService extends GenericService<HocSinh, Integer> {
+public interface HocSinhService extends GenericService<HocSinh, Long> {
     public HocSinh timHocSinhTheoEmail(String email) throws RemoteException;
 }
