+++ LTPT_UngDungThiTracNghiem_DHKTPM18C/src/main/java/service/impl/HocSinhServiceImpl.java
@@ -6,7 +6,7 @@ import service.HocSinhService;
 
 import java.rmi.RemoteException;
 
-public class HocSinhServiceImpl extends GenericServiceImpl<HocSinh, Integer> implements HocSinhService {
+public class HocSinhServiceImpl extends GenericServiceImpl<HocSinh, Long> implements HocSinhService {
     private HocSinhDAO hocSinhDAO;
     public HocSinhServiceImpl(HocSinhDAO hocSinhDAO) throws RemoteException {
         super(hocSinhDAO);
